# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Convention Cards Backend API
  version: 0.1.0
  description: Backend HTTP API for WBF-number login, player profiles, partnerships, convention cards, sharing, exports, and activity history.
servers:
  - url: http://localhost:3000
    description: Local API server
security:
  - bearerAuth: []
...
paths:
...
  /cards:
    get:
      tags: [Cards]
      summary: List convention cards owned by the signed-in player.
      parameters:
        - name: status
          in: query
          description: "Comma-separated card statuses. Example: draft,active."
          schema:
            type: string
        - name: includeArchived
          in: query
          description: Include archived cards when no status filter is supplied.
          schema:
            type: boolean
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Owned cards.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [cards]
                    properties:
                      cards:
                        type: array
                        items:
                          $ref: "#/components/schemas/ConventionCard"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Cards]
      summary: Create a blank draft card.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
      responses:
        "201":
          description: Draft created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/from-template:
    post:
      tags: [Cards]
      summary: Create a draft card from a template.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [templateSlug]
              properties:
                templateSlug:
                  type: string
                title:
                  type: string
                partnershipId:
                  type: [string, "null"]
                  format: uuid
      responses:
        "201":
          description: Draft created from template.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/reviews/pending:
    get:
      tags: [Cards]
      summary: List cards waiting for the signed-in partner's review.
      responses:
        "200":
          description: Pending partner review cards.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [cards]
                    properties:
                      cards:
                        type: array
                        items:
                          $ref: "#/components/schemas/ConventionCard"
        default:
          $ref: "#/components/responses/Error"
  /cards/{cardId}:
    parameters:
      - $ref: "#/components/parameters/CardId"
    get:
      tags: [Cards]
      summary: Load one owned card.
      responses:
        "200":
          description: Owned card.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags: [Cards]
      summary: Autosave editable draft fields.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardRequest"
      responses:
        "200":
          description: Draft updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/{cardId}/validation:
    parameters:
      - $ref: "#/components/parameters/CardId"
    get:
      tags: [Cards]
      summary: Validate an owned card for activation without changing it.
      responses:
        "200":
          description: Activation validation result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardValidationResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/{cardId}/revisions:
    parameters:
      - $ref: "#/components/parameters/CardId"
    post:
      tags: [Cards]
      summary: Create a draft revision from a rejected card.
      responses:
        "201":
          description: Revision created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/{cardId}/submit-for-approval:
    parameters:
      - $ref: "#/components/parameters/CardId"
    post:
      tags: [Cards]
      summary: Submit a draft for partner approval.
      responses:
        "200":
          description: Card submitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/{cardId}/review/approve:
    parameters:
      - $ref: "#/components/parameters/CardId"
    post:
      tags: [Cards]
      summary: Approve a submitted card as the invited partner.
      responses:
        "200":
          description: Card approved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/{cardId}/review/reject:
    parameters:
      - $ref: "#/components/parameters/CardId"
    post:
      tags: [Cards]
      summary: Reject a submitted card as the invited partner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rejectionReason:
                  type: string
                  maxLength: 1000
      responses:
        "200":
          description: Card rejected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/{cardId}/activate:
    parameters:
      - $ref: "#/components/parameters/CardId"
    post:
      tags: [Cards]
      summary: Activate a partner-approved card after validation.
      responses:
        "200":
          description: Card activated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/{cardId}/archive:
    parameters:
      - $ref: "#/components/parameters/CardId"
    post:
      tags: [Cards]
      summary: Archive a card using soft delete.
      responses:
        "200":
          description: Card archived.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/{cardId}/export:
    parameters:
      - $ref: "#/components/parameters/CardId"
    get:
      tags: [Cards]
      summary: Build a print-ready JSON export payload.
      responses:
        "200":
          description: Export payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardExportResponse"
        default:
          $ref: "#/components/responses/Error"
  /cards/{cardId}/history:
    parameters:
      - $ref: "#/components/parameters/CardId"
      - $ref: "#/components/parameters/Limit"
    get:
      tags: [Cards]
      summary: List recent activity events for an owned card.
      responses:
        "200":
          description: Card activity events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /shared/cards/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Sharing]
      security: []
      summary: Publicly load an active shared card by raw share token.
      responses:
        "200":
          description: Shared card.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSharedCardResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase access token
  parameters:
    CardId:
      name: cardId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    CardStatus:
      type: string
      enum: [draft, pending_partner_approval, partner_approved, partner_rejected, active, archived]
...
    ConventionCardData:
      type: object
      additionalProperties: true
    ConventionCard:
      type: object
      required: [id, ownerPlayerId, partnershipId, sourceCardId, revisionNumber, title, status, cardData, submittedAt, partnerReviewedByPlayerId, partnerReviewedAt, partnerRejectionReason, activatedAt, archivedAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        ownerPlayerId:
          type: string
          format: uuid
        partnershipId:
          type: [string, "null"]
          format: uuid
        sourceCardId:
          type: [string, "null"]
          format: uuid
        revisionNumber:
          type: integer
        title:
          type: string
        status:
          $ref: "#/components/schemas/CardStatus"
        cardData:
          $ref: "#/components/schemas/ConventionCardData"
        submittedAt:
          type: [string, "null"]
          format: date-time
        partnerReviewedByPlayerId:
          type: [string, "null"]
          format: uuid
        partnerReviewedAt:
          type: [string, "null"]
          format: date-time
        partnerRejectionReason:
          type: [string, "null"]
        activatedAt:
          type: [string, "null"]
          format: date-time
        archivedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ActivityEvent:
      type: object
      required: [id, eventType, actorPlayerId, entityType, entityId, cardId, partnershipId, shareLinkId, metadata, createdAt]
      properties:
        id:
          type: string
          format: uuid
        eventType:
          type: string
        actorPlayerId:
          type: [string, "null"]
          format: uuid
        entityType:
          type: string
          enum: [player, partnership, card, share_link]
        entityId:
          type: [string, "null"]
        cardId:
          type: [string, "null"]
          format: uuid
        partnershipId:
          type: [string, "null"]
          format: uuid
        shareLinkId:
          type: [string, "null"]
          format: uuid
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
...
    CardValidationIssue:
      type: object
      required: [code, path, message]
      properties:
        code:
          type: string
        path:
          type: string
        message:
          type: string
...
    CreateCardRequest:
      type: object
      properties:
        title:
          type: string
        partnershipId:
          type: [string, "null"]
          format: uuid
        cardData:
          $ref: "#/components/schemas/ConventionCardData"
    UpdateCardRequest:
      type: object
      properties:
        title:
          type: string
        cardData:
          $ref: "#/components/schemas/ConventionCardData"
...
    CardResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ConventionCard"
    CardValidationResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [validation]
          properties:
            validation:
              type: object
              required: [valid, issues]
              properties:
                valid:
                  type: boolean
                issues:
                  type: array
                  items:
                    $ref: "#/components/schemas/CardValidationIssue"
    CardExportResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: true
...
    PublicSharedCardResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [card, players, shareLink]
          properties:
            card:
              type: object
              required: [id, title, status, cardData, updatedAt]
              properties:
                id:
                  type: string
                title:
                  type: string
                status:
                  type: string
                  const: active
                cardData:
                  $ref: "#/components/schemas/ConventionCardData"
                updatedAt:
                  type: string
                  format: date-time
            players:
              type: object
              additionalProperties: true
            shareLink:
              type: object
              additionalProperties: true
    ActivityListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [events]
          properties:
            events:
              type: array
              items:
                $ref: "#/components/schemas/ActivityEvent"
...